Feature(live trader, connector): live trader, sends requests to alpaca using alpaca connector, basic error handling of donwloading yahoo data

+++ src/trading/util/connectors/base_connector.py
@@ -20,10 +20,10 @@ class BaseConnector():
         return None
 
     @abstractmethod
-    def place_buy_order(self, symbol: str, notional: int): pass
+    def place_buy_order(self, symbol: str, notional: float): pass
 
     @abstractmethod
-    def place_sell_order(self, symbol: str, qty: int): pass
+    def place_sell_order(self, symbol: str, qty: float): pass
 
     @abstractmethod
     def list_of_orders(self): pass
